requirements.txt
# Pinned production dependencies — regenerate with:
#   pip install pip-tools && pip-compile requirements.in -o requirements.txt
#
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
kombu==5.3.7
limits==3.13.0
Mako==1.3.5
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
redis==5.1.1
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
websockets==13.1
